+++ predict.py
@@ -8,7 +8,7 @@ import torch.nn.functional as F
 from PIL import Image
 from torchvision import transforms
 
-from data_loading import BasicDataset
+from utils.data_loading import BasicDataset
 from unet import UNet
 from utils import plot_img_and_mask
