+++ train.py
@@ -11,8 +11,8 @@ from torch import optim
 from torch.utils.data import DataLoader, random_split
 from tqdm import tqdm
 
-from data_loading import BasicDataset, CarvanaDataset
-from dice_score import dice_loss
+from utils.data_loading import BasicDataset, CarvanaDataset
+from utils.dice_score import dice_loss
 from evaluate import evaluate
 from unet import UNet
